requirements.txt
python-memcached==1.59
six==1.16.0
